Begin refactoring mailutils.py into mailutils.*
   * Start breaking mailutils.py into smaller files
   * Move some mailutils-style code from mailpile.search

Note: This commit leaves Mailpile in a non-working state, following
      commits are required to make sure things work.
